CBLDatabase.documentCount should explicitly return 64-bit int
A database can contain more than 4 billion documents, so NSUInteger
is the wrong property type (for 32-bit builds at least.)

+++ Objective-C/CBLDatabase.mm
@@ -354,7 +354,7 @@ static void dbObserverCallback(C4DatabaseObserver* obs, void* context) {
 #pragma mark - INTERNAL
 
 
-- (NSUInteger) documentCount {
+- (uint64_t) documentCount {
     return c4db_getDocumentCount(_c4db);
 }
